requirements.txt
asgiref==3.4.1
gunicorn==20.1.0
Django==3.2.7
django-decouple==2.1
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
requests~=2.26.0